requirements.txt
streamlit
google-generativeai
pyhton-dotenv
langchain
PyPDF2
chromadb